refactor: Lunchmenu 타입 폴더 구조 변경
